chore: remove uniq dist and binary manage of the klio binary
This patch improves the build system to put the following binaries in the Docker image:

- `/usr/bin/klio` 
- `/usr/bin/kopia`

Before these changes, the binaries were named after the architecture for which they were built.
The image can now be used without knowing the location of the Klio binary and directly relying on the entrypoint.