fix: append cluster name to GRPC requests
Kopia logins require user name and host in the `<username>@<hostname>`
format, allowing such users to reach the snapshots with their host name.

The same convention is being used by the GRPC WAL server, with the
hostname being set as the name of the PostgreSQL cluster. This patch
appends the hostname suffix to the user name in the GRPC client
connections.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>