ci: build operator image (EnterpriseDB/klio#94)
Add an operator Bakefile to allow using bake for image building.
Modify the operator Dockerfile to cross-compile the binary on the build-platform, instead
of using the possibly emulated target platform. This will reduce compilation time
when building for arm64.
Use UBI10 as a base for both operator and core images.
Add the relevant container building tasks for the operator to the Taskfile.
Change release-publish to also publish the operator image.

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Jonathan Battiato <jonathan.battiato@enterprisedb.com>