feat: add backup (EnterpriseDB/klio#90)
Avoid storing the CNPG infrastructure logging in the structs.
Use the Helm chart for the ephimeral development
environment instead of the Kustomize infrastracture.
Isolate the CNPGI plugin in a separate klio command, to
avoid `klio send-wal` to depend from a Kubernetes
environment to run correctly.
Keep track of instance name, WAL start/end file name.
Move cnpg out of the sendwal directory.
Use the correct cluster manifest with plugin in the docs.

---------

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Jonathan Battiato <jonathan.battiato@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Jonathan Battiato <jonathan.battiato@enterprisedb.com>