feat(send-wal): use a buffer when sending WALs (EnterpriseDB/klio#127)
This patch receives WAL blocks from PostgreSQL into a buffer whose
maximum size and flush deadline is configurable by the user.

PostgreSQL is using the same technique and allowing to set the WAL
buffer size. There is no deadline as the WAL files are flushed everytime
a transaction is committed.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>