ci: trigger and then skip ci on license changes (EnterpriseDB/klio#244)
We have a requirement to pass the CI runs, and this happens if the
required checks in GitHub are passed or skipped. Opening PRs without a
personal access token doesn't trigger a CI run, so the refresh-license
PRs could not be merged. Now we use the bot token, but we also want to
skip ci executions, since they're useless on PRs that work on licenses.

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>