fix: replication failure on timeline switch (EnterpriseDB/klio#224)
Fix a failure of the Klio WAL sender that happened if the Klio server
was a timeline behind. Starting LSN negotiation could negotiate an LSN
from a previous timeline that doesn't exist in the current one, which
could cause the klio wal sender to fail when asking that starting point
from PostgreSQL. Now we get the timeline from the Klio server, and
follow the timeline changes if necessary.

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>