feat: add opentelemetry metrics to core base and wal server (EnterpriseDB/klio#223)
Klio server is now able to export metrics and tracing for the base
and wal servers using OTEL.
Klio send-wal is now able to export tracing information.
The server API now includes metric configuration parameter.

OTEL can be configured on the instance manager by adding
environment variables and the relative projected volume (if needed)
in the Cluster configuration.

An example is provided in the samples.

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>