feat(auth): use mTLS and client-side certificate for authentication (EnterpriseDB/klio#366)
Instead of using a user name and a password to authenticate against a
Klio server, this patch makes Klio use a client certificate.

Support for the basic-auth authentication scheme has been removed and,
as a consequence, the same have happened for the htpasswd file.

The E2e tests have been changed to create a new CA issuer for each
Klio server and to generate the required client certificates.

Important: this patch requires the mTLS authentication support on
Kopia, that has been added on [1] and still not merged. For now this
works using Kopia as built from Leonardo's branch, which is available
at [2].

When the patch will be merged, we can remove the replacement in Klio's
`go.mod` and, more in general, we can migrate to the official Kopia
release.

[1]: https://github.com/kopia/kopia/pull/4886
[2]: https://github.com/leonardoce/kopia/tree/mtls

BREAKING CHANGE: previous configuration using htpasswd are no longer
supported.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>