refactor: use afero to abstract the file system (EnterpriseDB/klio#400)
This patch uses the afero in the WAL server, allowing it to use every
supported implementation of file system as persistent storage.

For now we use a composition of BasePathFs and OsFs to write the WAL
files to the local file system, preserving the existing behavior.
MemMapFs is used for the unit tests.

This refactoring work will be used when implementing the Tier 2, when
the WALs will be stored in a blob storage such as AWS S3.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>