refactor: separate repository from WAL server (EnterpriseDB/klio#401)
* refactor: separate repository from WAL server

This commit isolates all the code that work on the WAL repository in the
relative package. Before this, the feature was spread between the WAL
server with the "plumbing" features hosted in the "repository" package.

In addition to this, the "FS" field of the repository is not private,
isolating the user from the actual file system implementation.

This will be useful when implementing Tier 2, because we can now
directly using the repository without having to deal with a WAL server.

---------

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>