feat!: remove support for backupID from plugin configuration (EnterpriseDB/klio#425)
* feat!: remove support for backupID from plugin configuration

This patch removes the `backupID` field from the Plugin configuration,
preferring to use the recoveryTarget structure of CNPG.

When the `backupID` field in the `recoveryTarget` section is empty the
latest backup for the cluster is chosen. When a `recoveryTime` is
specified, the closest backup is chosen.

BREAKING CHANGE: the `backupID` field in the `PluginConfiguration`
object has been removed in favour of the
`cluster.spec.bootstrap.recovery.recoveryTarget.backupID` field.

---------

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>