feat: implement tier2 storage support with backup and WAL synchronization (EnterpriseDB/klio#412)
This implements a tier2 storage architecture for Klio, enabling automatic backup
and WAL file replication from tier1 (local) to tier2 (object store) storage.

NATS-based queues are used to coordinate backup and WAL synchronization
tasks as soon as a backup is completed or a WAL is received.

We rely on Kopia to synchronize the snapshots of the base backup on S3.
WALs are written on object store using afero-s3.

At this moment, only S3 is supported.

---------

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>