fix: clean up cache when initializing a new Kopia repository (EnterpriseDB/klio#483)
If a user drops the tier1 data PVC without dropping the corresponding cache, we
initialize a new Kopia repository but we still use the cache of the old one.

To prevent that and to ensure the Kopia repository is correctly paired up with
the cache, we now drop the cache every time a repository is initialied.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Signed-off-by: John Long <john.long@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: John Long <john.long@enterprisedb.com>