fix: avoid using snapshot IDs in backup metadata (EnterpriseDB/klio#492)
We currently reference the snapshot IDs of tablespaces and of pgdata in the
snapshot of pg_controldata. This allows Klio to use them together consistently.

When migrating snapshots to tier2 via kopia snapshot migrate, snapshots get a
different ID - thus broking the link.

This patch avoids doing that and use the backup name tag. By doing that, we
avoid using multiple Kopia internals and we can just rely on its CLI.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>