refactor: isolate ImageConfiguration and TLSConfiguration in the API (EnterpriseDB/klio#537)
This PR isolates the `image`, `imagePullPolicy` and `pullSecret`
fields of the Server API into a structure named `ImageConfiguration`
which has been embeddede into the spec stanza.

The same happened to the `tlsSecretName` and to the
`clientCASecretName` fields, which became the `TLSConfiguration`
structures.

Those structures will be later reused to create the `RecoverySource`
API.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>